Expose view helpers for focused testing
